pyproject.toml
[project]
name = "ev-charging-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "comet-ml>=3.55.0",
    "gymnasium>=1.2.2",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
